update changes and news for release
